loainfo,loawa 추가
