+++ deploy-commands.js
@@ -6,10 +6,16 @@ const { clientId, guildIds, token } = require('./config.json');
 
 const commands = [];
 const commandFiles = fs.readdirSync('./commands').filter(file => file.endsWith('.js'));
+const loaCommandFiles = fs.readdirSync('./commands/lostark').filter(file => file.endsWith('.js'));
 
 for (const file of commandFiles) {
 	const command = require(`./commands/${file}`);
 	commands.push(command.data.toJSON());
+}
+
+for (const file of loaCommandFiles) {
+	const loaCommand = require(`./commands/lostark/${file}`);
+	commands.push(loaCommand.data.toJSON());
 }
 
 const rest = new REST({ version: '9' }).setToken(token);
